Moved SVG Icon components from Navbar to icon folder in root.
